Add rmsecrets option to parseConfigWithSecrets() and activate ts-check
Parameter table of parseConfigWithSecrets() was reconstructed, for one
parameter `encoding` was put into a new parameter `options` which is an
object with a few optional keys. Another option `rmsecrets` is put into
the object to choose whether to hide the `secrets` key of the returning
value.

The script has also activated `//@ts-check` for type checking like that
in TypeScript. But these scripts may rewrite to TypeScript files for
full code experience and strict and simple type annotations.