Re-initialize all code for type checking
This commit has many changed files for adding type checking support to
all JavaScript files. The stuff had tried using pure TypeScript before
but it caused many unexpected errors and numerous times of refactoring.
So we still use JavaScript for core features with ".d.ts" files for type
annotations. Note that the configuration files for this goal may be
changed in the future to meet more demands.