Add package @types/mysql
The aim of adding this package manually is to support TypeScript typing
in VS Code.
We can see that package mysql has been added to the project, and
JavaScript files can parse `require('mysql')` successfully and link to
the corresponding typing declarations in Microsoft's TypeScript work
folder (`%APPDATA%/Microsoft/TypeScript/4.5/node_modules/@types/mysql`),
but it fails when it comes to custom `.d.ts` files.
So we suppose that JS in VS Code can correctly link to the "global"
typing declarations but TS can only search them in the local folder.