Update database utility as an archive
It's a pity that the database utility needs a little big update on its
supported databases. We found that Redis is very lite but doesn't
support some features of an object-relational database which may have
a boost on our development. So we decided to archive all the supports
on Redis and replace it with MySQL. Actually, MySQL can "define" a
transaction that can easily roll back if some steps fail. This feature
makes the "message queues" (in the future?) more convenient.