Append "Sync" to database function names
We found that it's important for database I/O operations to replace
synchronous functions with asynchronous ones. These old functions may be
unable to handle with returns and other variables. They are still kept
in the code tree, but supposed to be deprecated and not recommended.
Please await asynchronous functions to solve these problems.