Add `@typescript-eslint/no-explicit-any` rule
This rule is still regarded as a warning like default settings, but it
won't be triggered when it comes to a rest argument (`...args`).

In TypeScript, anything is assignable to `unknown` type, but `unknown` is
not assignable to anything but itself and any without a type assertion or
a control flow based narrowing. So the rest argument (annotated
`unknown[]`) of the function `inst.connectSync()` of `MySQLConn` will
cause an ESlint error that `unknown` (one element of the rest argument)
can't be assignable to arguments of `mysql.Connection.connect` (whose
rest argument is annotated `any[]`). If the declarations are modified and
support `any` type, the problem will be solved.

You can find more information on https://stackoverflow.com/a/51439876/12002560.