Change `_gethash()` to simple MD5
This function for database operations is to provide a primary key for
tables to avoid duplication of entries by mistake. The original version
was implemented by simple concatenation of all the string-type columns.
But now the length of one column has been longer, so it's a waste of
space for the old one. The new version uses `crypto-js/md5` to calculate
MD5 hash of the previous concatenation, fixing the length of `_hashval`
column to acceptable 32.