Fix the bug of wrong types of `req.query.compact` and `.no_peer_id`
These two query parameters seem to be 0 or 1. But for the HTTP server
both of them are regarded as a string which can be "0" or "1". If they
were directly passed to the validation before, it would think that
they were not included in `[0, 1]` and made this validation fail. Now
these two will be converted to `Number` first (strings with invalid
characters will be converted to `NaN`), and the validation will work
well.