Add two fields to `ActiveClientsConn` and refine code
Two fields "ip" (string) and "port" (number) have been added to the
database of active clients. It's obvious that `ip`s and `port`s of the
clients are very important for a tracker.

With more fields than before, code of handling with these fields would
become longer and less elegant if we still used the old coding style.
Now we try to take advantage of anonymous functions and loops to face
potential challenges of more fields of this and other databases.