Fix the bug of unexpected errors thrown from rejected promises
`Promise.reject()` can create a rejected promise easily. It is simple to
catch the rejected promise from some chained ones via `promise.catch()`.
But an error will be thrown which makes the program exit if it is a
return value of one promise.

There is a `ReferenceError` and two `TypeError`s inside the code. They
have been removed in this commit.