Remove "left" field from data row hash factors
In the previous versions of `_gethash()` in `utils/common/database.js`,
it included all fields of ActiveClients database except `_hashval` for
hash factors. But we found that "left" field is changing when down-
loading and this value in the database should be synchronously changed.
So the field cannot be one of the factors of `_gethash()`.