fix: use client-side configs instead of server-side
In previous code `routes/announce.js` and `utils/announce/process.js`
used server-side configurations (`cfg.server`) for database connections.
This is a mistake due to many reasons. The database library is stored in
this repo, actually being used by many other scripts in other repos.
When in this repo it's very easy to get the server-side configurations.
But when it comes to other repos that's really hard. For another point,
these "server-side" configurations are for creating servers, while
"client-side" for connecting to the servers. In the two changed files
the library was used to communicate with the database server, so they
*should* use the client-side ones.