chore: add results of queries to returns of update functions
In the previous code the update functions didn't expose the results of
their query operations, only those of *UPDATE* (or *INSERT*). Sometimes
we would like to *SELECT* and *UPDATE*/*INSERT* at the same time, so we
added this feature.

But it's obvious that it may result in more consumption of resources,
for there may be more duplicated database I/O than before. This will be
solved in the future.