feat: add pre-initialization support
Now it is available to use `PREINIT` environment variable to pre-
initialize the databases before handling the HTTP server (just by
setting it to "true", "TRUE", "1", etc.).

Note that the asynchronous functions and database I/O may make the
HTTP server start listening before the database initialization. But
in many cases that's OK.