chore: avoid completes, seeders and leechers become negative
If the requests of database I/O are not so frequent, we are convinced no
such kind of errors will be generated. To make the service keeps going,
we decide to add fallback values for `completes`, `seeders` and
`leecheres` fields of new torrent rows. They will be zero when they
reach negative after some calculations.