fix: router of `/announce` returns with appropriate headers
In issue #5 we found that our responses didn't have appropriate headers.
For the headers we only set `Content-Type` to `text/plain`, which is
defined in BitTorrent specifications. The potential bug hadn't been
revealed until some recent tests were performed. The data returned by
the `/announce` router was mixed with some special characters, which are
necessary for data transferring (i.e. ending marks), showed in the issue
above. This kind of mixture confused some BitTorrent clients (i.e.
qBittorrent).

In this commit we added more `Content-Type` and return Buffer of the
bencoded data. The Buffer will be converted in `binary` encoding to
avoid encoding errors (in previous responses the data was encoded to
UTF-8, not ISO-8859-1 (`binary`) we need).

Close #5